Make ImageClassification API an ITrainerEstimator and refactor code. (#4372)
* Make ImageClassification API an ITrainerEstimator and refactor code.

* comments.

* comments.

* Fix premature session close.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.

* PR feedback.
