+++ docs/samples/Microsoft.ML.Samples.GPU/Microsoft.ML.Samples.GPU.csproj
@@ -11,7 +11,7 @@
 
   <ItemGroup>
     <Compile Include="..\Microsoft.ML.Samples\Program.cs" Link="Program.cs" />
-    <Compile Include="..\Microsoft.ML.Samples\Dynamic\ImageClassification\*.cs">
+    <Compile Include="..\Microsoft.ML.Samples\Dynamic\Trainers\MulticlassClassification\ImageClassification\*.cs">
       <Link>Dynamic\ImageClassification\%(FileName)</Link>
     </Compile>
     <Compile Include="..\Microsoft.ML.Samples\Dynamic\TensorFlow\*.cs">
