+++ src/Microsoft.ML.Data/Scorers/PredictedLabelScorerBase.cs
@@ -42,8 +42,8 @@ namespace Microsoft.ML.Data
             private readonly AnnotationUtils.AnnotationGetter<ReadOnlyMemory<char>> _getScoreValueKind;
             private readonly DataViewSchema.Annotations _predColMetadata;
             private BindingsImpl(DataViewSchema input, ISchemaBoundRowMapper mapper, string suffix, string scoreColumnKind,
-                bool user, int scoreColIndex, DataViewType predColType)
-                : base(input, mapper, suffix, user, DefaultColumnNames.PredictedLabel)
+                bool user, int scoreColIndex, DataViewType predColType, string predictedLabelColumnName = DefaultColumnNames.PredictedLabel)
+                : base(input, mapper, suffix, user, predictedLabelColumnName)
             {
                 Contracts.AssertNonEmpty(scoreColumnKind);
                 Contracts.Assert(DerivedColumnCount == 1);
@@ -82,7 +82,7 @@ namespace Microsoft.ML.Data
             }
 
             public static BindingsImpl Create(DataViewSchema input, ISchemaBoundRowMapper mapper, string suffix,
-                string scoreColKind, int scoreColIndex, DataViewType predColType)
+                string scoreColKind, int scoreColIndex, DataViewType predColType, string predictedLabelColumnName = DefaultColumnNames.PredictedLabel)
             {
                 Contracts.AssertValue(input);
                 Contracts.AssertValue(mapper);
@@ -90,7 +90,7 @@ namespace Microsoft.ML.Data
                 Contracts.AssertNonEmpty(scoreColKind);
 
                 return new BindingsImpl(input, mapper, suffix, scoreColKind, true,
-                    scoreColIndex, predColType);
+                    scoreColIndex, predColType, predictedLabelColumnName);
             }
 
             public BindingsImpl ApplyToSchema(DataViewSchema input, ISchemaBindableMapper bindable, IHostEnvironment env)
@@ -272,7 +272,7 @@ namespace Microsoft.ML.Data
         [BestFriend]
         private protected PredictedLabelScorerBase(ScorerArgumentsBase args, IHostEnvironment env, IDataView data,
             ISchemaBoundMapper mapper, RoleMappedSchema trainSchema, string registrationName, string scoreColKind, string scoreColName,
-            Func<DataViewType, bool> outputTypeMatches, Func<DataViewType, ISchemaBoundRowMapper, DataViewType> getPredColType)
+            Func<DataViewType, bool> outputTypeMatches, Func<DataViewType, ISchemaBoundRowMapper, DataViewType> getPredColType, string predictedLabelColumnName = DefaultColumnNames.PredictedLabel)
             : base(env, data, registrationName, Contracts.CheckRef(mapper, nameof(mapper)).Bindable)
         {
             Host.CheckValue(args, nameof(args));
@@ -292,7 +292,7 @@ namespace Microsoft.ML.Data
             Host.Check(outputTypeMatches(scoreType), "Unexpected predictor output type");
             var predColType = getPredColType(scoreType, rowMapper);
 
-            Bindings = BindingsImpl.Create(data.Schema, rowMapper, args.Suffix, scoreColKind, scoreColIndex, predColType);
+            Bindings = BindingsImpl.Create(data.Schema, rowMapper, args.Suffix, scoreColKind, scoreColIndex, predColType, predictedLabelColumnName);
             OutputSchema = Bindings.AsSchema;
         }
