+++ src/Microsoft.ML.Dnn/DnnRetrainTransform.cs
@@ -11,13 +11,13 @@ using System.Text;
 using Microsoft.ML;
 using Microsoft.ML.CommandLine;
 using Microsoft.ML.Data;
+using Microsoft.ML.Dnn;
 using Microsoft.ML.Internal.Utilities;
 using Microsoft.ML.Runtime;
 using Microsoft.ML.Transforms;
-using Microsoft.ML.Transforms.Dnn;
 using NumSharp;
 using Tensorflow;
-using static Microsoft.ML.Transforms.Dnn.DnnUtils;
+using static Microsoft.ML.Dnn.DnnUtils;
 using static Tensorflow.Binding;
 
 [assembly: LoadableClass(DnnRetrainTransformer.Summary, typeof(IDataTransform), typeof(DnnRetrainTransformer),
