+++ src/Microsoft.ML.Dnn/DnnUtils.cs
@@ -13,9 +13,10 @@ using System.Security.Principal;
 using Microsoft.ML.Data;
 using Microsoft.ML.Runtime;
 using Tensorflow;
+using static Microsoft.ML.Dnn.ImageClassificationTrainer;
 using static Tensorflow.Binding;
 
-namespace Microsoft.ML.Transforms.Dnn
+namespace Microsoft.ML.Dnn
 {
     internal static class DnnUtils
     {
@@ -260,25 +261,15 @@ namespace Microsoft.ML.Transforms.Dnn
         /// <param name="modelPath">The model to load.</param>
         /// <param name="metaGraph"></param>
         /// <returns></returns>
-        internal static DnnModel LoadDnnModel(IHostEnvironment env, string modelPath, bool metaGraph = false)
-        {
-            var session = GetSession(env, modelPath, metaGraph);
-            return new DnnModel(env, session, modelPath);
-        }
+        internal static DnnModel LoadDnnModel(IHostEnvironment env, string modelPath, bool metaGraph = false) =>
+            new DnnModel(GetSession(env, modelPath, metaGraph), modelPath);
 
-        /// <summary>
-        /// Load TensorFlow model into memory.
-        /// </summary>
-        /// <param name="env">The environment to use.</param>
-        /// <param name="arch">The architecture of the model to load.</param>
-        /// <param name="metaGraph"></param>
-        /// <returns></returns>
-        internal static DnnModel LoadDnnModel(IHostEnvironment env, ImageClassificationEstimator.Architecture arch, bool metaGraph = false)
+        internal static DnnModel LoadDnnModel(IHostEnvironment env, Architecture arch, bool metaGraph = false)
         {
-            var modelPath = ImageClassificationEstimator.ModelLocation[arch];
+            var modelPath = ModelLocation[arch];
             if (!File.Exists(modelPath))
             {
-                if (arch == ImageClassificationEstimator.Architecture.InceptionV3)
+                if (arch == Architecture.InceptionV3)
                 {
                     var baseGitPath = @"https://raw.githubusercontent.com/SciSharp/TensorFlow.NET/master/graph/InceptionV3.meta";
                     using (WebClient client = new WebClient())
@@ -293,7 +284,7 @@ namespace Microsoft.ML.Transforms.Dnn
                         ZipFile.ExtractToDirectory(Path.Combine(Directory.GetCurrentDirectory(), @"tfhub_modules.zip"), @"tfhub_modules");
                     }
                 }
-                else if (arch == ImageClassificationEstimator.Architecture.ResnetV2101)
+                else if (arch == Architecture.ResnetV2101)
                 {
                     var baseGitPath = @"https://aka.ms/mlnet-resources/image/ResNet101Tensorflow/resnet_v2_101_299.meta";
                     using (WebClient client = new WebClient())
@@ -301,7 +292,7 @@ namespace Microsoft.ML.Transforms.Dnn
                         client.DownloadFile(new Uri($"{baseGitPath}"), @"resnet_v2_101_299.meta");
                     }
                 }
-                else if (arch == ImageClassificationEstimator.Architecture.MobilenetV2)
+                else if (arch == Architecture.MobilenetV2)
                 {
                     var baseGitPath = @"https://tlcresources.blob.core.windows.net/image/MobileNetV2TensorFlow/mobilenet_v2.meta";
                     using (WebClient client = new WebClient())
@@ -309,7 +300,7 @@ namespace Microsoft.ML.Transforms.Dnn
                         client.DownloadFile(new Uri($"{baseGitPath}"), @"mobilenet_v2.meta");
                     }
                 }
-                else if (arch == ImageClassificationEstimator.Architecture.ResnetV250)
+                else if (arch == Architecture.ResnetV250)
                 {
                     var baseGitPath = @"https://tlcresources.blob.core.windows.net/image/ResNetV250TensorFlow/resnet_v2_50_299.meta";
                     using (WebClient client = new WebClient())
@@ -320,8 +311,7 @@ namespace Microsoft.ML.Transforms.Dnn
 
             }
 
-            var session = GetSession(env, modelPath, metaGraph);
-            return new DnnModel(env, session, modelPath);
+            return new DnnModel(GetSession(env, modelPath, metaGraph), modelPath);
         }
 
         internal static Session GetSession(IHostEnvironment env, string modelPath, bool metaGraph = false)
