+++ src/Microsoft.ML.Dnn/LearningRateScheduler.cs
@@ -1,8 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
-namespace Microsoft.ML.Transforms
+using System;
+using System.Collections.Generic;
+
+namespace Microsoft.ML.Dnn
 {
     /// <summary>
     /// A class that contains the current train state to use for learning rate scheduling.
